refactor: connections layouts
@khanhhaduc I can’t see the data cos I don’t have connections on my
local. Here’s the basic stuff.

TODO:

1) SettingsIntegrationsDetail: Ii haven’t got classes yet (hopefully
next commit) but I’ve refactored the top layout.

2) InvoiceSection: I’ve added in the components, but you’ll have to
hook up the data. Hopefully it’ll be obvious as it’s the same as the
hacky version.

3) Entity Section: I added in the button group. You’ll want to make it
go to the right place.

4: ConnectionSetting: I’ve added in a couple of links BUT it’s erroring
out on me, so I’ve had to comment out the actual ${currentOrg} bits. As
before, there’s no actual styling on this yet, just layout.